+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Unreleased
+
+* Deprecated `Hex::to_array3` in favor of `Hex::to_cubic_array`
+
 ## 0.4.2
 
 * Fixed `conversion` module was private (#42)